requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.3
asyncpg==0.29.0
python-dotenv==1.0.1
